LLM-Course-Assignments-2025/01-NLP/submissions/rtca-rag-system/requirements.txt
# Core
python>=3.9
pyyaml>=6.0
pydantic>=2.0
pandas>=2.0
numpy>=1.24

# Deep Learning
torch>=2.0
transformers>=4.35
accelerate>=0.24
peft>=0.6
bitsandbytes>=0.41  # For quantization

# Vector DB & Retrieval
faiss-cpu>=1.7
chromadb>=0.4
langchain>=0.0.340
rank_bm25>=0.2
sentence-transformers>=2.2

# Document Processing
pdfplumber>=0.10
python-docx>=1.1
pymupdf>=1.23
markdown>=3.5
tiktoken>=0.5

# Web Frameworks
gradio>=4.0
streamlit>=1.28
fastapi>=0.104
uvicorn>=0.24
httpx>=0.25
jinja2>=3.1

# Evaluation
evaluate>=0.4
rouge-score>=0.1
sacrebleu>=2.3
scikit-learn>=1.3

# Testing
pytest>=7.4
pytest-cov>=4.1
pytest-asyncio>=0.21

# Logging & Monitoring
loguru>=0.7
prometheus-client>=0.19
structlog>=23.0

# Caching
redis>=5.0
pymemcache>=4.0

# Development Tools
black>=23.0
flake8>=6.0
mypy>=1.5
pre-commit>=3.4
jupyter>=1.0
ipython>=8.0

# Other
tqdm>=4.65
colorama>=0.4
rich>=13.5
click>=8.1
tenacity>=8.2